% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_stats.R
\name{check_args}
\alias{check_args}
\title{Check validity of arguments for statistics}
\usage{
check_args(object, trait, head_var, weight, taxon_levels, merge_to)
}
\description{
Check validity of arguments in functions \code{\link[=trait_stats]{trait_stats()}} and
\code{\link[=trait_proportion]{trait_proportion()}}.
}
\keyword{internal}
