% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverconvert-class.R
\docType{class}
\name{coverconvert}
\alias{coverconvert}
\alias{coverconvert-class}
\title{Cover conversion tables}
\description{
Cover conversion tables for \linkS4class{vegtable} objects.

This class implements conversions from different cover scales in percentage
cover. For transformations to percentage cover, the function \code{\link[=transform]{transform()}}
should be than used.
}
\section{Slots}{

\describe{
\item{\code{value}}{List containing the levels of each scale.}

\item{\code{conversion}}{List with the respective start and end cut levels for the
scale levels.}
}}

\examples{
showClass("coverconvert")

## Add a custom scale
Scale <- new("coverconvert")
Scale$my_scale <- list(
	 value=factor(c("low","medium","high"), levels=c("low","medium","high")),
	 conversion=c(0,50,75,100))
summary(Scale)

}
\seealso{
\code{\link[=tv2coverconvert]{tv2coverconvert()}} \link{braun_blanquet}.
}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}
}
