% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers2samples.R
\name{layers2samples}
\alias{layers2samples}
\alias{layers2samples,vegtable,character,character-method}
\alias{layers2samples,vegtable,character,missing-method}
\title{Add information from slot 'layers' into slot 'samples'}
\usage{
layers2samples(object, layer, variable, ...)

\S4method{layers2samples}{vegtable,character,character}(object, layer, variable, ...)

\S4method{layers2samples}{vegtable,character,missing}(object, layer, variable, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{vegtable}.}

\item{layer}{Character value indicating a target layer.}

\item{variable}{Character vector with the names of variables to be inserted
in slot \code{samples}.}

\item{...}{Further arguments to be passed among methods.}
}
\value{
An object of class \linkS4class{vegtable} with variables added to samples.
}
\description{
Slot layers may include additional information that should be moved to
samples in order to use it by \code{\link[vegtable:subset]{vegtable::subset()}}, \code{\link[vegtable:aggregate]{vegtable::aggregate()}}
or \code{\link[vegtable:crosstable]{vegtable::crosstable()}} methods.

If names of variables are not provided, all variables from the respective
layer table will be inserted in slot \code{samples}.
}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}.
}
