% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_names.R
\name{match_names}
\alias{match_names}
\alias{match_names,character,vegtable-method}
\title{Search matchings between character and taxlist objects.}
\usage{
\S4method{match_names}{character,vegtable}(x, object, ...)
}
\arguments{
\item{x}{A character vector with names to be compared.}

\item{object}{An object of class \linkS4class{vegtable} to be compared with.}

\item{...}{Further arguments passed to \code{\link[taxlist:match_names]{taxlist::match_names()}}.}
}
\description{
Names provided in a character vector will be compared with names stored in
slot \code{taxonNames} of an object of class \linkS4class{taxlist} by using the
function \code{\link[stringdist:stringsim]{stringdist::stringsim()}}.

This method is applied to the slot \code{species} in the input \linkS4class{vegtable}
object.
}
\seealso{
\code{\link[taxlist:match_names]{taxlist::match_names()}} \code{\link[stringdist:stringsim]{stringdist::stringsim()}}
}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}
}
