% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{rewrite_formulas}
\alias{rewrite_formulas}
\title{Re-writing formulas for print output}
\usage{
rewrite_formulas(shaker, companion)
}
\arguments{
\item{shaker}{A \linkS4class{shaker} object.}

\item{companion}{A companion data set which is either missing or a
\linkS4class{vegtable} object.}
}
\value{
A formated output text.
}
\description{
Mediating between syntax and print format.
}
\keyword{internal}
