% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_det.R
\name{update_det}
\alias{update_det}
\alias{update_det,vegtable,character-method}
\alias{update_det,vegtable-method}
\title{Update by determined specimens}
\usage{
update_det(x, specimens, ...)

\S4method{update_det}{vegtable,character}(x, specimens, ...)
}
\arguments{
\item{x}{A \linkS4class{vegtable} object to be updated.}

\item{specimens}{A character vector indicating the names of tables included
in slot \strong{layers} with updates to be applied.
Note that they will be applied in the same order of the vector in the
case of multiple updates.}

\item{...}{Further arguments (not yet in use).}
}
\description{
Reference specimens can be integrated in slot \strong{layers} within a
\linkS4class{vegtable} object.
Updated entries in the specimens can be updated in slot \strong{samples} by using
this function.
Alternatively expert opinions can be inserted and applied in case of
disagreement with the original records.
}
