% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_layers-data.R
\docType{data}
\name{veg_layers-data}
\alias{veg_layers-data}
\alias{veg_layers}
\title{Information on vegetation layers}
\format{
A data frame.
}
\usage{
veg_layers
}
\description{
A data frame with a standard classification on vegetation layers.
}
\examples{
summary(veg_layers)
}
\keyword{datasets}
