% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegtable-deprecated.R
\name{vegtable-deprecated}
\alias{vegtable-deprecated}
\alias{match_names,character,vegtable-method}
\alias{merge_taxa,vegtable,numeric,missing-method}
\alias{vegtable2kml}
\title{Deprecated functions in package \pkg{vegtable}}
\usage{
\S4method{match_names}{character,vegtable}(x, object, ...)

\S4method{merge_taxa}{vegtable,numeric,missing}(object, concepts, level, ...)

vegtable2kml(obj, ...)
}
\description{
The functions listed below are deprecated and will be defuncti in the near
future.
}
\section{\code{match_names}}{

For \code{match_names(object, ...)}, use \code{match_names(object@species, ...)}
}

\section{\code{merge_taxa}}{

For \code{merge_taxa(object, ...)}, use \code{merge_taxa(object@species, ...)}
}

\section{\code{vegtable2kml}}{

Use leaflet to map plot observations.
}

\keyword{internal}
