% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspect_con-data.R
\docType{data}
\name{aspect_conv-data}
\alias{aspect_conv-data}
\alias{aspect_conv}
\title{Conversion of aspect classes to azimuth}
\format{
A numeric vector of values in degrees for the symbols used as names.
}
\usage{
aspect_conv
}
\description{
Conversion table required to transform values of aspect to azimuth in
degrees.
}
\examples{
aspect_conv[c("N", "S", "ENE", "SSW")]
}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}
}
\keyword{datasets}
