% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braun_blanquet-data.R
\docType{data}
\name{braun_blanquet-data}
\alias{braun_blanquet-data}
\alias{braun_blanquet}
\title{Conversion of Braun-Blanquet codes to cover percentage

Cover values conversion as \linkS4class{coverconvert} object.

Object of class \linkS4class{coverconvert} contains conversion tables
usually from a categorical variable (a cover scale) to a numerical one
(equivalent percentage cover value).
Cover values are stored as range for each level in the scale (minimum and
maximum cover value).}
\format{
An object of class \code{\linkS4class{coverconvert}}.
}
\usage{
braun_blanquet
}
\description{
Conversion of Braun-Blanquet codes to cover percentage

Cover values conversion as \linkS4class{coverconvert} object.

Object of class \linkS4class{coverconvert} contains conversion tables
usually from a categorical variable (a cover scale) to a numerical one
(equivalent percentage cover value).
Cover values are stored as range for each level in the scale (minimum and
maximum cover value).
}
\examples{
names(braun_blanquet)
summary(braun_blanquet)
summary(braun_blanquet$b_bbds)
}
\seealso{
\linkS4class{coverconvert} \code{\link[=cover_trans]{cover_trans()}}
}
\keyword{datasets}
