\name{braun_blanquet-data}
\docType{data}
\alias{braun_blanquet-data}
\alias{braun_blanquet}

\title{Conversion of Braun-Blanquet codes to cover percentage.}
\description{
Data set as \code{\linkS4class{coverconvert}} object.
}
\usage{
data(braun_blanquet)
}
\format{
An object of class \code{\linkS4class{coverconvert}}.
}
\examples{
library(vegtable)
data(braun_blanquet)
}
\keyword{datasets}
