\name{clean}
\docType{methods}
\alias{clean}
\alias{clean,vegtable-method}

\title{Clean orphaned records in vegtable object.}
\description{
Delete entries in slots \code{header} and \code{species} orphaned by
modifications in slot \code{samples}.
}
\usage{
\S4method{clean}{vegtable}(object, ...)
}
\arguments{
\item{object}{a \code{\linkS4class{vegtable}} object.}
\item{...}{further arguments passed from or to other methods.}
}
\details{
This function delete orphaned records usually generated by modifications in the
slot \code{taxonRelations}.
This function should be applied to optimize the allocated size of a
\code{\link[=taxlist-class]{taxlist}} object and to assure its validity.
}
\value{
A clean \code{\linkS4class{taxlist}} object.
}
\author{
Miguel Alvarez (\email{kamapu78@gmail.com}).
}
\examples{
\donttest{
library(vegtable)
data(sweadataveg)

## Direct manipulation of slot samples generates an invalid object
sweadataveg@samples <- sweadataveg@samples[1:500,]
summary(sweadataveg)

## Now apply cleaning
sweadataveg <- clean(sweadataveg)
summary(sweadataveg)

Easplist <- clean(Easplist)
summary(Easplist)
}
}
\keyword{methods}
