\name{clean}
\docType{methods}
\alias{clean}
\alias{clean,vegtable-method}

\title{Clean orphaned records in vegtable object.}
\description{
Delete entries in slots \code{header} and \code{species} orphaned by
modifications in slot \code{samples}.
}
\usage{
\S4method{clean}{vegtable}(object, ...)
}
\arguments{
\item{object}{a \code{\linkS4class{vegtable}} object.}
\item{...}{further arguments passed from or to other methods.}
}
\details{
This function delete orphaned records usually generated by modifications in the
slot \code{taxonRelations}.
This function should be applied to optimize the allocated size of a
\code{\link[=taxlist-class]{taxlist}} object and to assure its validity.
}
\value{
A clean \code{\link[=taxlist-class]{taxlist}} object.
}
\author{
Miguel Alvarez (\email{kamapu78@gmail.com}).
}
\examples{
library(taxlist)
data(Easplist)
\donttest{
## Direct manipulation of slot taxonRelations generates an invalid object
Easplist@taxonRelations <- Easplist@taxonRelations[1:5,]
summary(Easplist)
}
## Now apply cleaning
Easplist <- clean(Easplist)
summary(Easplist)
}
\keyword{methods}
