% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as}
\alias{as}
\alias{as.list,vegtable-method}
\alias{coerce,vegtable,list-method}
\alias{coerce,coverconvert,list-method}
\alias{as<-}
\alias{as<-,vegtable-method}
\alias{as.list,coverconvert-method}
\alias{as<-,coverconvert-method}
\title{Coerce objects to lists}
\usage{
\S4method{as.list}{vegtable}(x, ...)

as(from) <- value

\S4method{as}{vegtable}(from) <- value

\S4method{as.list}{coverconvert}(x, ...)

\S4method{as}{coverconvert}(from) <- value
}
\arguments{
\item{x, from}{An object to be coerced.}

\item{...}{further arguments passed from or to other methods.}

\item{value}{A character value indicating the class to be coerced to. This is
only required by the replacement methods.}
}
\value{
An object of class \code{list}.
}
\description{
Coerce vegtable objects to a list with every slot as a component of the list.
This way allows to explore content and solve problems when validity checks
fail.

Coercion is applied for different classes by vegtable.
}
\examples{
## vegtable as list
veg <- as(Kenya_veg, "list")
names(veg)

## coverconvert as list
as(Kenya_veg@coverconvert, "list")

}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}
}
