\name{coverconvert-class}
\docType{class}
\alias{coverconvert-class}
\alias{coverconvert}

\title{Class 'coverconvert'.}
\description{
Descriptioin
}
\section{slots}{
\describe{
\item{value}{list containing the levels of each scale.}
\item{conversion}{list with the respective start, end and cut levels for the
	respective scale classes.}
}
}
\details{
More details.
}
\author{
Miguel Alvarez (\email{kamapu78@gmail.com}).
}
\seealso{
\code{\link{tv2coverconvert}}.
}
\examples{
library(vegtable)

showClass("coverconvert")
}
