% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dune_veg-data.R
\docType{data}
\name{dune_veg-data}
\alias{dune_veg-data}
\alias{dune_veg}
\title{Dutch dune meadows as vegtable

Data set from the package \link[vegan:vegan]{vegan::vegan}, converted to a \linkS4class{vegtable}
object.}
<<<<<<< HEAD
\format{An object of class \code{\linkS4class{vegtable}}.}
=======
\format{
An object of class \code{\linkS4class{vegtable}}.
}
>>>>>>> refs/remotes/origin/miguel
\source{
Original data were imported from \link[vegan:dune]{vegan::dune}.
}
\usage{
dune_veg
}
\description{
Dutch dune meadows as vegtable

Data set from the package \link[vegan:vegan]{vegan::vegan}, converted to a \linkS4class{vegtable}
object.
}
\examples{
summary(dune_veg)
}
\references{
\bold{Jongman RHG, ter Braak CJF, van Tongeren OFR (1987).}
\emph{Data analysis in community and landscape ecology.} Pudoc, Wageningen,
NL.
}
\keyword{datasets}
