% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{header}
\alias{header}
\alias{header,vegtable-method}
\alias{header<-}
\alias{header<-,vegtable,data.frame-method}
\title{Retrieve or replace slot header in vegtable objects}
\usage{
header(x, ...)

\S4method{header}{vegtable}(x, ...)

header(x) <- value

\S4method{header}{vegtable,data.frame}(x) <- value
}
\arguments{
\item{x}{Object of class \linkS4class{vegtable}.}

\item{...}{Further arguments passed to or from other methods.}

\item{value}{Data frame to be set as slot \code{header}.}
}
\description{
Retrieve or replace the content of slot \code{header} in \linkS4class{vegtable} objects.
}
\examples{
head(header(Kenya_veg))

}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}
}
