% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{names}
\alias{names}
\alias{names,vegtable-method}
\alias{names<-,vegtable-method}
\alias{names<-}
\alias{dimnames,vegtable-method}
\alias{dimnames}
\alias{names,coverconvert-method}
\alias{names<-,coverconvert-method}
\title{Retrieve names of vegtable and coverconvert objects}
\usage{
\S4method{names}{vegtable}(x)

\S4method{names}{vegtable}(x) <- value

\S4method{dimnames}{vegtable}(x)

\S4method{names}{coverconvert}(x)

\S4method{names}{coverconvert}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{coverconvert} or \linkS4class{vegtable}.}

\item{value}{A character vector used for replacement methods.}
}
\value{
Either a vector or a list (in the case of \code{dimnames()}) with
the names of variables.
}
\description{
Quick access to column names in slot header and names of conversion codes.

These methods provide a quick display of the contents in
\linkS4class{coverconvert} and \linkS4class{vegtable} objects.
}
\examples{
names(Kenya_veg@coverconvert)
names(Kenya_veg)
dimnames(Kenya_veg)

}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}.
}
