% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_layer.R
\name{new_layer}
\alias{new_layer}
\alias{new_layer,vegtable,data.frame-method}
\alias{new_layer<-}
\alias{new_layer<-,vegtable,data.frame-method}
\title{Insert new classification of vegetation layers}
\usage{
new_layer(object, layer, ...)

\S4method{new_layer}{vegtable,data.frame}(object, layer, ...)

new_layer(object, ...) <- value

\S4method{new_layer}{vegtable,data.frame}(object, ...) <- value
}
\arguments{
\item{object}{A \linkS4class{vegtable} object.}

\item{layer, value}{A data frame including information on vegetation layers.}

\item{...}{Further arguments passed among methods.}
}
\value{
A \linkS4class{vegtable} object with the inserted new relation.
}
\description{
A new information table for vegetation layers.
}
\examples{
## Modify name in samples
names(Kenya_veg@samples) <- replace_x(names(Kenya_veg@samples),
    old = "LAYER", new = "layer")

## Add installed data frame to layers
new_layer(Kenya_veg) <- veg_layers

## Take a look in the result
summary(Kenya_veg@samples$layer)
}
