\name{set_group}
\docType{methods}
\alias{set_group}
\alias{set_group,shaker,taxlist,character-method}
\alias{set_group,shaker,vegtable,character-method}
\alias{set_pseudo}
\alias{set_pseudo,shaker,taxlist,character-method}
\alias{set_pseudo,shaker,vegtable,character-method}

\title{Set slots in shaker objects.}
\description{
Properly set of slots in \code{\linkS4class{shaker}} objects.
}
\usage{
\S4method{set_group}{shaker,taxlist,character}(shaker, content, group,
	group_id, ...)

\S4method{set_group}{shaker,vegtable,character}(shaker, content, group, ...)

\S4method{set_pseudo}{shaker,taxlist,character}(shaker, content, pseudo,
	pseudo_id, ...)

\S4method{set_pseudo}{shaker,vegtable,character}(shaker, content, pseudo, ...)
}
\arguments{
\item{shaker}{Describe.}
\item{content}{Describe.}
\item{group,pseudo}{Describe.}
\item{group_id,pseudo_id}{Describe.}
\item{...}{Describe.}
}
\details{}
\value{}
\author{
Miguel Alvarez (\email{kamapu78@gmail.com}).
}
\seealso{
\code{\link{}}
}
\examples{
## Add example for this function
}
