% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa2samples.R
\name{taxa2samples}
\alias{taxa2samples}
\alias{taxa2samples.vegtable}
\alias{taxa2samples,vegtable-method}
\title{Insert taxonomic information into slot samples}
\usage{
taxa2samples(object, ...)

\method{taxa2samples}{vegtable}(
  object,
  merge_to,
  include_levels,
  add_relations = FALSE,
  add_traits = FALSE,
  ...
)
}
\arguments{
\item{object}{Object of class \linkS4class{vegtable}.}

\item{...}{Further arguments passed to \code{\link[taxlist:merge_taxa]{taxlist::merge_taxa()}}.}

\item{merge_to}{Character value indicating the level to which the taxa
have to be merged.}

\item{include_levels}{Character vector indicating the levels to be considered
in the output object. It can be used to exclude some taxonomic ranks.}

\item{add_relations}{A logical value indicating whether the content of slot
\strong{taxonRelations} have to be inserted in slot \strong{samples} or not.}

\item{add_traits}{A logical value indicating whether the content of slot
\strong{taxonTraits} have to be inserted in slot \strong{samples} or not.}
}
\value{
An object of class \linkS4class{vegtable}.
}
\description{
For some statistical purposes it may be necessary to insert taxonomic
information into the slot samples.

This function is also used internally by functions such as
\link[=count_taxa]{vegtable::count_taxa()}.
}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}
}
