% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegtable-class.R
\docType{class}
\name{vegtable-class}
\alias{vegtable-class}
\alias{vegtable}
\title{Class vegtable.}
\description{
Class holding vegetation-plot data sets.
Designed to content all information stored in \strong{Turboveg} databases in
just one object.

This class was designed to include information of relevés, header data and
species in just one object. Objects can be created by calls of the form
\code{new("vegtable", ...)}.
}
\section{Slots}{

\describe{
\item{\code{description}}{A named character vector containing metadata.}

\item{\code{samples}}{A data frame with samples list.}

\item{\code{header}}{A data frame with plots data.}

\item{\code{species}}{Species list as a \linkS4class{taxlist} object.}

\item{\code{layers}}{A list including strata within samples as data frames.}

\item{\code{relations}}{A list including popup lists as data frames.}

\item{\code{coverconvert}}{A scale conversion object of class \linkS4class{coverconvert}.}

\item{\code{syntax}}{A list including syntaxonomic lists either as data frames or as
\linkS4class{taxlist} objects.}
}}

\examples{
showClass("vegtable")
}
\seealso{
\code{\link[=tv2vegtable]{tv2vegtable()}}
}
