% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegtable-deprecated.R
\name{vegtable-deprecated}
\alias{vegtable-deprecated}
\title{Deprecated functions in package \pkg{vegtable}}
\description{
The functions listed below are deprecated and will be defuncti in the near
future.
\itemize{
\item \code{\link{vegtable2kml}}: This function is deprecated, and will
be removed in the next version of this package.
}
}
\keyword{internal}
