% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_taxa.R, R/vegtable2kml.R,
%   R/match_names.R, R/vegtable-deprecated.R
\name{vegtable-deprecated}
\alias{vegtable-deprecated}
\alias{merge_taxa}
\alias{vegtable2kml}
\alias{match_names,character,vegtable-method}
\title{Merge concepts}
\usage{
merge_taxa(object, ...)

vegtable2kml(obj, ...)

\S4method{match_names}{character,vegtable}(x, object, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{vegtable} to be compared with.}

\item{...}{Further arguments passed to \code{\link[taxlist:match_names]{taxlist::match_names()}}.}

\item{obj}{Input object containing coordinate values.}

\item{x}{A character vector with names to be compared.}

\item{concepts}{Numeric (integer) vector including taxon concepts to be
merged.}

\item{level}{Character value indicating the level to which the taxa
have to be merged.}

\item{file}{Character value with the name of output file (including file
extension).}

\item{coords}{Either a character vector or a formula indicating the names of
coordinate values.}

\item{srs}{Spatial reference system as \code{proj4string}.}
}
\value{
An object of class \linkS4class{vegtable}.

A KML file, which will be automatically opened in \strong{Google Earth}.
}
\description{
Merge taxon concepts form into single ones or insert accepted names to slot
samples.

This method is applied to a function defined in the package \link{taxlist-package}
and only modify the slot \code{species} in the input \code{object}.

The use of \code{taxa2samples()} with \code{merge_to} argument will produce a similar
result as using \code{merge_taxa} with \code{level} argument, but \code{taxa2samples()}
will replace the records in slot samples by the respective accepted names
without any modification in slot species.
Additionally taxon concept IDs will be addes as columns in samples and taxon
traits if indicated in argument \code{add_traits}.

This function is producing and displaying KML files.

Georeferenced plots can be quickly displayed in
\href{https://www.google.com/intl/en_us/earth/}{Google Earth} using this function.

Names provided in a character vector will be compared with names stored in
slot \code{taxonNames} of an object of class \linkS4class{taxlist} by using the
function \code{\link[stringdist:stringsim]{stringdist::stringsim()}}.

This method is applied to the slot \code{species} in the input \linkS4class{vegtable}
object.

The functions listed below are deprecated and will be defuncti in the near
future.
}
\section{\code{merge_taxa}}{

For \code{merge_taxa(object, ...)}, use \code{merge_taxa(object@species, ...)}
}

\section{\code{match_names}}{

For \code{match_names(object, ...)}, use \code{match_names(object@species, ...)}
}

\seealso{
\link{vegtable-deprecated}

\link{vegtable-deprecated}
}
\author{
Miguel Alvarez \email{kamapu@posteo.de}
}
\keyword{internal}
