\name{vegtable_stat}
\alias{vegtable_stat}

\title{General Statistics from 'vegtable' Data Sets.}

\description{
This function calculates general statistics of local TURBOVEG databases as
required by GIVD (Global Index of Vegetation-Plot Databases,
\url{http://botanik3.botanik.uni-greifswald.de/givd/}).
}
\usage{
vegtable_stat(vegtable)
}
\arguments{
\item{vegtable}{an object of class \code{\link[=vegtable-class]{vegtable}}.}
}
\details{
This function is based on a script delivered by GIVD for
summarising statistics required in the descriptions of databases (see meta data
in the page of the Global Index for Vegetation-Plot Databases).
}
\author{GIVD. Adapted by Miguel Alvarez.}
\examples{
## Same example as in 'Africa_veg' data
}
