% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wetlands-data.R
\docType{data}
\name{Wetlands-data}
\alias{Wetlands-data}
\alias{Wetlands}
\alias{Wetlands_veg}
\title{Vegetation-plots from Tanzania}
\format{
An object of class \linkS4class{shaker} (\code{Wetlands}) and the respective companion
as \linkS4class{vegtable} object (\code{Wetlands_veg}).
}
\source{
\url{http://www.givd.info/ID/AF-00-006}.
}
\usage{
Wetlands
}
\description{
A subset of \url{http://www.givd.info/ID/AF-00-006SWEA-Dataveg} with plots
sampled in Tanzania.
}
\examples{
summary(Wetlands)
summary(Wetlands_veg)

}
\references{
\bold{Alvarez M (2017).} Classification of aquatic and
semi-aquatic vegetation in two East African sites: Cocktail definitions and
syntaxonomy. \emph{Phytocoenologia}.
}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}
}
\keyword{datasets}
